Big update

+++ ProShopping/src/main/java/code/proshopping/createAccountController.java
@@ -30,49 +30,54 @@ public class createAccountController implements Initializable {
     public void initialize(URL url, ResourceBundle resourceBundle) {
         genderBox.setItems(listGender);
     }
-
-    public void cancelAction(ActionEvent event){
-
-    }
     public void okAction(ActionEvent event){
 
-        if(!passwordText.getText().equals(confirmPasswordText.getText())){
+        if(usernameText.getText().isEmpty() || passwordText.getText().isEmpty() || age.getText().isEmpty() || confirmPasswordText.getText().isEmpty()){
             Alert alert = new Alert(Alert.AlertType.INFORMATION);
             alert.setTitle("Notification");
-            alert.setContentText("Password confirmation doesn't match!");
-            alert.showAndWait();
-            return;
-        }
-
-        File fileAccCheck = new File("src/Data/account/customer/" + usernameText.getText() + ".txt");
-        if (fileAccCheck.exists()){
-            Alert alert = new Alert(Alert.AlertType.INFORMATION);
-            alert.setTitle("Notification");
-            alert.setContentText("Account already exists!");
+            alert.setContentText("Please fill in all information!");
             alert.showAndWait();
         }
-        else {
-            try{
-                BufferedWriter bw = new BufferedWriter(new FileWriter("src/Data/account/customer/" + usernameText.getText() + ".txt"));
-                bw.write(passwordText.getText());
-                bw.newLine();
-                bw.write("Information: " + age.getText() + "/" + genderBox.getValue());
-                bw.newLine();
-                bw.write("Balance: 0");
-                bw.newLine();
-                bw.write("Cart: ");
-                bw.newLine();
-                bw.write("Shopping history:");
-                bw.newLine();
-                bw.close();
-            } catch (IOException e) {
-                throw new RuntimeException(e);
+        else{
+            if(!passwordText.getText().equals(confirmPasswordText.getText())){
+                Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                alert.setTitle("Notification");
+                alert.setContentText("Password confirmation doesn't match!");
+                alert.showAndWait();
+                return;
             }
 
-            Alert alert = new Alert(Alert.AlertType.INFORMATION);
-            alert.setTitle("Notification");
-            alert.setContentText("Account created successfully!");
-            alert.showAndWait();
+            File fileAccCheck = new File("src/Data/account/customer/" + usernameText.getText() + ".txt");
+            if (fileAccCheck.exists()){
+                Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                alert.setTitle("Notification");
+                alert.setContentText("Account already exists!");
+                alert.showAndWait();
+            }
+            else {
+                try{
+                    BufferedWriter bw = new BufferedWriter(new FileWriter("src/Data/account/customer/" + usernameText.getText() + ".txt"));
+                    bw.write(passwordText.getText());
+                    bw.newLine();
+                    bw.write("Information: " + age.getText() + "/" + genderBox.getValue());
+                    bw.newLine();
+                    bw.write("Balance: 0");
+                    bw.newLine();
+                    bw.write("Cart: ");
+                    bw.newLine();
+                    bw.write("Shopping history:");
+                    bw.newLine();
+                    bw.close();
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                }
+
+                Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                alert.setTitle("Notification");
+                alert.setContentText("Account created successfully!");
+                alert.showAndWait();
+            }
         }
+
     }
 }
